Add files via upload
In this Internship Application Assignment with the Application Software Centre at IIT Bombay, the project entails two main components: backend development using Django and frontend development using ReactJS. The goal is to create a comprehensive web-based system for listing and managing courses offered by an educational institution.

On the backend, a Django-based RESTful API will be developed to handle course-related operations, including course creation, deletion, and retrieval, as well as the management of course delivery instances. This component will serve as the core data and logic layer for the project.

Concurrently, the frontend application will be built using ReactJS to provide an intuitive user interface for interacting with the courses. It will enable users to view, create, and delete courses, as well as access detailed information about individual courses and their respective instances.

Version control will be maintained through GitHub repositories for both the backend and frontend, with regular commits and code updates. Additionally, Dockerization of the applications is a crucial aspect.